requirements.txt
selenium>=4.0.0
undetected-chromedriver>=3.5.0
webdriver-manager>=4.0.0
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=9.0.0
Unidecode>=1.3.0
python-dotenv>=1.0.0
